+++ src/service/impl_http/mod.rs
@@ -1,0 +1,4 @@
+#[cfg(feature = "default-http-client")]
+pub mod client;
+pub mod evaluate;
+pub mod factory;
